# docs/deapi-mirror/articles/openapi.hono.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dryAPI Gateway API",
    "description": "Unified OpenAI-compatible inference gateway with async job orchestration, provider controls, and webhook tooling.",
    "version": "2026-03-15"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin worker URL"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API key",
        "description": "Use `Authorization: Bearer <api-key>`."
      }
    }
  },
  "paths": {
...
    "/v1/queue/batch-scaling/stream": {
      "get": {
        "operationId": "streamQueueBatchScaling",
        "tags": [
          "Queue"
        ],
        "summary": "Stream live queue batch scaling updates",
        "description": "Streams server-sent events with dynamic batch scaling payloads built from queue metrics and runtime analytics.",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "SSE stream that emits live queue batch scaling updates."
          }
        }
      }
    },
...
  }
}